Extract Flair to shared components
And refactor import
