+++ src/ui/components/Flair/index.ts
@@ -1,0 +1,1 @@
+export { default as Flair } from "./Flair.svelte";
